+++ src/features/blockchain/components/Blockchain.Provider.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
 
-import { BlockchainContextValue } from '../@types'
 import { BlockchainContextProvider } from '../contexts'
+import { BlockchainContextValue } from '../types'
 
 export const BlockchainProvider = React.memo(function BlockchainProvider({
   children,