+++ src/features/blockchain/types/enums.ts
@@ -1,3 +1,6 @@
+/**
+ * @deprecated use `BlockchainNamespace` instead
+ */
 export enum SupportedBlockchainNamespace {
   EIP_155 = 'eip155',
   NEAR = 'near',